+++ Changelog.md
@@ -7,7 +7,8 @@
 - Move Stack Trashed into Better Trash
 - Added Trash the Trash to Better Trash Slot
 - Added Deposit Middle Click
-- Added Complete Quick Stack
+- Added Complete Quick Stack to Better Quick Stack
+- Added Limited Personal Quick Stack to Better Quick Stack
 - Fixed Quick with Frame Skip Off
 
 ## v0.8.2